refactor queries
