+++ src/app/modules/search/components/SearchBox.js
@@ -1,9 +1,12 @@
 import React, { Component } from 'react';
+import { connect} from 'react-redux';
 import Tooltip from 'rc-tooltip';
-import Loader from "../../../components/Misc/Loader";
-import {Queries} from "../../../components/index";
+import SkyLight from 'react-skylight';
+import SketchPicker from 'react-color';
+import {Query} from "../../../components/index";
+import {editSearch} from "../actions";
 
-export default class SearchBox extends Component {
+class SearchBox extends Component {
     constructor(props) {
         super(props);
 
@@ -11,7 +14,8 @@ export default class SearchBox extends Component {
 
         this.state = {
             query: '',
-            selectValue: this.props.indexes[0]
+            selectValue: this.props.indexes[0],
+            editSearchValue: null
         };
     }
 
@@ -32,9 +36,27 @@ export default class SearchBox extends Component {
         this.setState({query: e.target.value});
     }
 
+    handleEditSearch(search) {
+        this.setState({editSearchValue: search});
+        this.refs.editDialog.show();
+    }
+
+    handleChangeQueryColorComplete(color) {
+        const { dispatch } = this.props;
+        const search = Object.assign({}, this.state.editSearchValue);
+
+        search.color = color.hex;
+
+        dispatch(editSearch(search.q, {
+            color: color.hex
+        }));
+
+        this.setState({editSearchValue: search});
+    }
+
     render() {
-        const { connected, enabled } = this.props;
-        const { query } = this.state;
+        const { connected, enabled, searches } = this.props;
+        const { query, editSearchValue } = this.state;
 
         let tooltipStyles = {};
 
@@ -42,36 +64,64 @@ export default class SearchBox extends Component {
             tooltipStyles.visibility = 'hidden';
         }
 
+        const editQueryDialogStyles = {
+            backgroundColor: '#fff',
+            color: '#000',
+            width: '400px',
+            height: '400px',
+            marginTop: '-200px',
+            marginLeft: '-200px',
+        };
+
+        let edit_query = null;
+        if (editSearchValue) {
+            edit_query = <form>
+                <SketchPicker
+                    color={ editSearchValue.color }
+                    onChangeComplete={(color) => this.handleChangeQueryColorComplete(color) }
+                />
+                <span className="colorBall" style={{backgroundColor: editSearchValue.color}}/>
+                { `${editSearchValue.q} (${editSearchValue.items.length})` }
+            </form>;
+        }
+
         return (
             <nav id="searchbox" className="[ navbar ][ navbar-bootsnipp animate ] row" role="navigation">
-                <div className="logo-container">
+                <div className="logoContainer">
                     <img className={`logo ${connected ? 'connected' : 'not-connected'}`} src="/images/logo.png" title={connected ? "Marija is connected to the backendservice" : "No connection to Marija backend available" } />
                 </div>
-                <div className="queries-container">
-                    <Queries />
-                </div>
-                <div className="search-container">
-                    <div className="form-group">
-                        <form onSubmit={this.handleSubmit.bind(this)}>
-                            <Tooltip
-                                overlay="Select at least one datasource and one field"
-                                placement="bottomLeft"
-                                overlayStyle={tooltipStyles}
-                                arrowContent={<div className="rc-tooltip-arrow-inner" />}>
+                <div className="queriesContainer">
+                    {searches.map(search => <Query search={search} key={search.q} handleEdit={() => this.handleEditSearch(search)}/>)}
 
-                                <input
-                                    className="form-control"
-                                    placeholder="Search something"
-                                    value={ query }
-                                    onChange={this.handleQueryChange.bind(this)}
-                                    disabled={ !enabled }
-                                />
-
-                            </Tooltip>
-                        </form>
-                    </div>
+                    <form onSubmit={this.handleSubmit.bind(this)}>
+                        <Tooltip
+                            overlay="Select at least one datasource and one field"
+                            placement="bottomLeft"
+                            overlayStyle={tooltipStyles}
+                            arrowContent={<div className="rc-tooltip-arrow-inner" />}>
+                            <input
+                                className="searchInput"
+                                placeholder="Search"
+                                value={ query }
+                                onChange={this.handleQueryChange.bind(this)}
+                                disabled={ !enabled }
+                            />
+                        </Tooltip>
+                    </form>
                 </div>
+                <SkyLight dialogStyles={editQueryDialogStyles} hideOnOverlayClicked ref="editDialog" title="Update query" >
+                    { edit_query }
+                </SkyLight>
             </nav>
         );
     }
 }
+
+const select = (state, ownProps) => {
+    return {
+        ...ownProps,
+        searches: state.entries.searches
+    };
+};
+
+export default connect(select)(SearchBox);
