+++ src/scss/app.scss
@@ -35,18 +35,7 @@ body {
         box-shadow: 0 1px 0 0 rgba(0,0,0,0.25);
 
         nav{
-            padding: 10px 0px;
-
-            .form-group{
-                input.form-control{
-                    border: 0px solid transparent !important;
-                    background-color: transparent !important;
-                    background-image: url('../images/search-icon.png');
-                    background-repeat: no-repeat;
-                    background-size: 26px 25px;
-                    background-position: left center;
-                }
-            }
+            padding: 10px 0 0;
         }
 
     }
@@ -82,20 +71,19 @@ body {
         cursor: default;
     }
 
-    .logo-container{
+    .logoContainer{
         width: 81px;
-        height: 68px;
         margin: -15px 0px;
         box-shadow: inset -1px 0 0 0 rgba(0,0,0,0.20);
-        text-align:center;
-        float:left;
+        text-align: center;
+        flex-shrink: 0;
     }
 
     .logo {
         max-width: auto;
         max-height: 42px;
         margin-left: 15px;
-        padding: 13px 0px;
+        padding: 11px 0px;
         
         &.not-connected{
             animation: fadeIn 1s infinite alternate;
