+++ src/scss/panes.scss
@@ -2,6 +2,7 @@
     background-color: $pane-background-color;
     margin: 0;
     box-shadow: -1px 0 0 0 rgba(0,0,0,0.15);
+    z-index: 2;
 
     .pane-header {
         position: relative;
